backend/requirements.txt
Flask
Flask-CORS
python-dotenv
requests
waitress